pyproject.toml
[tool.poetry]
name = "applemusicrp"
version = "3.2.1"
description = "Discord Rich Presence for Apple Music w/ Album Artwork!"
authors = ["wxllow <willow@wxllow.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
dialite = "^0.5.3"
coverpy = "^0.0.4"
appdirs = "^1.4.4"
pypresence = "^4.2.1"
toml = "^0.10.2"
psutil = {version = "^5.9.1", platform = "windows"}
pywin32 = {version = "^304", platform = "windows"}
pystray = {version = "^0.19.3", platform = "windows"}
Pillow = {version = "^9.1.1", platform = "windows"}
rumps = {version = "^0.3.0", platform = "darwin"}
setuptools = "^62.6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
briefcase = "^0.3.8"
pre-commit = "^2.20.0"
py2exe = {version="^0.12.0", platform = "windows"}

[tool.briefcase]
project_name = "applemusicrp"
bundle = "dev.wxllow.dev"
version = "3.2.1"
url = "https://github.com/wxllow/applemusicrp"
license = "MIT"
author = 'wxllow'
author_email = "willow@wxllow.dev"

[tool.briefcase.app.applemusicrp]
formal_name = "AppleMusicRP"
description = "Discord Rich Presence for Apple Music w/ Album Artwork!"
icon = "logos"
sources = ['src/applemusicrp']
requires = [
    'dialite',
    'coverpy',
    'appdirs',
    'pypresence',
    'toml',
    'rich',
    'setuptools'
]

[tool.briefcase.app.applemusicrp.macOS]
requires = [
    'rumps',
]
